employee_management_system/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
passlib[bcrypt]
jose
